refactor(svar-2): extract /proc monitoring into monitor.rs
Moves the ~130-line Linux CPU sampler out of lib.rs; documents it as Linux-only
and prints n/a (not a misleading 0%) for CPU columns on macOS where /proc is absent.
